+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+### Fixed
+
+- Resolve correct project from GitHub PR URLs when multiple projects share the same upstream remote.
 ## [0.1.0] - 2025-02-15
 
 ### Added